restructured code for better maintainability
